requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
filelock==3.9.0
Flask==2.3.3
fonttools==4.42.1
frozenlist==1.4.0
gunicorn==20.0.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.2
matplotlib==3.7.2
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numpy==1.24.1
openai==0.27.9
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
Pillow==9.3.0
psutil==5.9.5
py-cpuinfo==9.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.28.1
scipy==1.11.2
seaborn==0.12.2
six==1.16.0
sympy==1.11.1
-r torch_requirements.txt
tqdm==4.66.1
typing_extensions==4.4.0
tzdata==2023.3
ultralytics==8.0.163
urllib3==1.26.13
Werkzeug==2.3.7
yarl==1.9.2
